requirements.txt
requests
transformers>=4.35.0
torch>=2.0.0
openai
pyyaml
streamlit>=1.28.0